obi.asd minor fix to path stuff We had unsatisfiable classes which are resolved by edits to external.owl These included: organism,viruses,bacteria,archaea,eukaryotes (plus some collateral damage)
Here's why:

Organism: equivalent viruses or bacteria or archaea or eukaryotes
siblings (because they were only subclass of material)
viruses
archaea
bacteria
eukaryotes

Disjoints then add (because that are the same level)
organism disjoint viruses
organism disjoint archaea
organism disjoint bacteria
organism disjoint eukaryotes

Organism is therefore the union of 4 things it is disjoint with. That leaves it empty.
Because it is *equivalent* to the union, each of them have to be empty too.

The repair was, in external.owl, to make viruses, bacteria, archaea,
eukaryotes subclass of organism, rather than material.



git-svn-id: http://svn.code.sf.net/p/obi/code@583 a3e8fd31-a726-0410-8195-c3f0c8a19530