Completion of long overdue action on my part. Added the following
Mass spectrometer
has_part ion_source
has_part mass_analyzer
has_part ion_detector

Ion_source: An ion source is a device that is part of a mass
spectrometer that ionizes the material under analysis. The ions are
then transported by magnetic or electric fields to the mass analyzer.
Techniques for ionization have been key to determining what types of
samples can be analyzed by mass spectrometry. Electron ionization and
chemical ionization are used for gases and vapors. In chemical
ionization sources, the materila is ionized by chemical ion-molecule
reactions during collisions in the source. Two techniques often used
with liquid and solid biological samples include electrospray
ionization (due to John Fenn PMID 2675315.) and matrix-assisted laser
desorption/ionization (MALDI, due to M. Karas and F. Hillenkamp
(Measuring Mass: From Positive Rays to Proteins by Michael A. Grayson
(Editor) (ISBN 0-941901-31-9))).
http://en.wikipedia.org/wiki/Mass_spectrometry#Ion_source

Mass_analyzer: A Mass analyzer is a device that separates ions
according to their mass-to-charge ratio. All mass spectrometers are
based on dynamics of charged particles in electric and magnetic fields
in vacuum where the following two laws of Lorentz force law and
Newton's second law of motion apply
http://en.wikipedia.org/wiki/Mass_spectrometry#Mass_analyzer

Ion_detector: An ion detector is a device that measures and records
the charge induced or current produced when an ion passes by or hits a
surface. In a scanning instrument the signal produced in the detector
during the course of the scan versus where the instrument is in the
scan (at what m/Q) will produce a mass spectrum, a record of ions as a
function of m/Q.
http://en.wikipedia.org/wiki/Mass_spectrometry#Detector


-Ontology is consistent

git-svn-id: http://svn.code.sf.net/p/obi/code@700 a3e8fd31-a726-0410-8195-c3f0c8a19530