minor corrections - preparing for resubmission
1. Changes:
For the OWLED paper, please put into the acknowledgements all the following grants (not just the NuGO, as previously indicated by me):
the BBSRC (BB/D524283/1, BB/E025080/1), the EU Network of Excellence NuGO (NoE 503630), the EU Carcinogenomics (PL037712)

2. Typos:

- page 4
Other options are possible, for instance having the 
script that generates additional information or using software that extracts a module of 
the external ontology. 

- page 5
The reports are rated according what action needs to be taken: simple warnings for 
those errors that can be corrected automatically by script

- page 6 
footnote is duplicated
 The latest version of OBI is available at http://purl.obofoundry.org/obo/obi.owl 

- page 9
An ontology interface that supported fluidly moving between 
the component trees, the defined classes, an

- page 9
Our solution for disjoints is not entirely satisfactory. Declaring disjoint policy for 
whole trees where the siblings are all mutually disjoint is appealing, but there are 
exceptions., Co



git-svn-id: http://svn.code.sf.net/p/obi/code@990 a3e8fd31-a726-0410-8195-c3f0c8a19530