Fix punning bug by declaring class BFO:0000003
Protege 5 complained about punning BFO:0000062 and BFO:0000063
as both ObjectProperty and AnnotationProperty.
The fix seems to be asserting that BFO:0000003 is an owl:Class
in external-by-hand.owl.
Otherwise BFO:0000003 is assumed to be an AnnotationProperty,
and since it is the domain and range of those relations
HermiT infers that the relations are AnnotationProperties.
(I think.)


git-svn-id: http://svn.code.sf.net/p/obi/code@4065 a3e8fd31-a726-0410-8195-c3f0c8a19530